local_client_requirements.txt
requests
tqdm
torch
torchaudio
soundfile
silero
google-auth
